conflicting command resolved
